Create timed WakeLock and use it to handle notifications processing
Also creates a NotificationsAndroidService, woken up by an alarm, to help coordinate with other devices that might also be reading messages. When we detect another device has just read a message, we only silently show notifications and after some time, wake up the service to make sure that the notifications are displayed with sound again.

The timed wakelock is to ensure that we keep the CPU running while processing and displaying notificatons while the app is in the background