Drop dependency on SAN list check only for hostname verification.
Changed domain check to not rely only on SAN list, as this may not always be
present.
This prevents MITM attack in case where attacker was able to install malicious certificate
on client device or compromised some Certificate Authority. In that case attacker
would be able to create certificate with unexpected SAN list and thus circumvent
our certificate pinning.
This will not be possible anymore.