Disable autowiring by default for final source signals.
SourceSignals usually don't need autowiring (unless used by subclass).
Disabling it by default prevent unnecessary warnings when `currentValue` is used.

Warning log in `currentValue` is there for a reason, it may be dangerous to use currentValue with unwired signals,
we want devs to pay attention to that logs, as those are potentials bugs.