Fix WebSocketService ping alarms.
Changed restartIntent to use proper flag.
When using ONE_SHOT alarm manager is only able to use the intent once, meaning that it won't be a recurring array as we would expect.
This way service was never woken up, so websocket ping was not performed if cpu was not active.
Also app would not be restarted after process is killed.