Ensure that cookie and token are fetched at the point of sending request
If several requests from the ZnetClient came at once, then the AuthenticationManager would load the token for each one at the beginning. If one of the requests first updated the token from /access, then the following requests wouldn't update the token they use and so several unecessary requests to  could follow. Now the token is only loaded if a request to  is actually going to be made, and then the token is saved, all in sequence to prevent further race conditions.

Furthermore, a new Cookie type is introduced to make the logic a bit clearer as to when a cookie is actually defined or not, and to simplify logging