Include BE drift when saving lastUiVisibleTime
There can be large amounts of drift from the timestamps on each of the message events compared to local time (one example was 1m 7s). This means, after minimising the app, we might miss any notifications received in this time. Including the BE time is a quick fix, but we should probably update events at the source instead of trying to remember to use the drift in all different parts of the app.