Update MessageIndexStorage to extend CachedStorage.
Added caching to MessageIndexStorage.
Added explicit prefetching of all search results in `searchText`, this makes
search slightly slower but ensures that later calls to `get..` are fast.
This is needed for `getNormalizedContentForMessage` method, which is
called from each view in results list, we need to ensure that this method is
fast, it should always read data from cache, never go to DB.
We now have a hard limit on number of search results, so this prefetching is
safe, will load at most 1024 entries, this is small enough to be quite fast.