Remove messages cursor order calculation.
Checking cursor ordering is not really necessary since we know its order when MessagesCursor is created.
We can avoid doing this hacky check using a simple parameter. It's always good
to get rid of unnecessary code.

Also, cursor was using binary search in IndexWindow, this search was
incorrectly assuming ascending order, as order check was not used there.