Ensure that lifecycle state does not stop listening to GSM call state
There could be a slight delay between the call-state changing to OFFHOOK and us receiving an idle lifecycle state. Since the device should be alive if the user accepts a GSM call, we shouldn't have to worry about locks and can freely close calling in the background as the user accepts. Before, we might not be notified about the state change in time, and a wire-call would be left open at the same time as the GSM call is answered