fixes ImageLoaderSpec, removes previews loading for bitmaps
Unit now ImageLoader tried to download the same bitmap twice - once as a preview and once as the full image and if the full image was downloaded first, it cancelled the preview download. This commit removes previews download.
Also, it adds a test for animated gifs previews download and a small utils
class with one method which returns the stack trace as a string. It can be used in debug messages.