Move GcmGlobal and GcmService into one class
It's unnecessary to delete tokens for different user registrations. We only really need to register a push token on first login on a given device (at which point we can also more reliably check that play services is available and up to date) and then we can keep that token for all future logins.