Remove push token from account on logout
If a new user logs in and registers with the same global token, BE will also delete the registration for any previous users using that token. That means, if we log in again, we need to re-register the first user with the same global token. Setting their login token to None will force registration on login again.