[AN-5144] Use randombytes from libsodium for keys.
Using libsodium was recommended instead of SecureRandom due to security
concerns.
Added small jni wrapper exposing needed method and changed AESUtils to use it.

Note: this commit includes binary version of libsodium.so, but this file is
only used during compilation, it's not included in released aar.
Cryptobox already includes latest version of libsodium and we want to use
that one.