Ensure conv order and unarchiving events are handled in batch, sequentially, before conv state handling
This should ensure that all conversations are updated with their last event time and unarchived if necessary, all at once. This should reduce the number of changes to the UI conversation list. Furthermore, a signal is introduced so the UI can block its signals while these ordering events are being handled.