Remove MessagesList class from public API.
Removed no longer used API, it was only used in integration tests, had to
update many of them.
This api was relying on current ZMessaging instance, which will go away soon,
it's better to get rid of those calsses instead of trying to make it work with
concurrent zms instances.