Introduce Either for definition of teamId in account service
Having either allows us to represent the state an account can be in before we know if it's a team or not. This is important for performing operations accross multiple accounts or in logging in, when we need to check if the account is asscociated with a team and be sure that it's actually defined.