Fix migration of old private accounts
Previous accounts would be set with an undefined team status, preventing us from creating a ZMessaging instance for them. The team data would never be updated, and the account is then unusable.
Also, if we load an account from the database without a defined team/personal account status, trigger the loading of the team data again.