Update account data correctly in registration
Ensures that each step of the ensureFullyRegistered method is responsible for updating the account storage completely before passing onto the next step - there was a bug where one step expected it to be done, but the account was only updated in memory, leading to client registration state loss and ultimately registered multiple clients per login.

This commit also tidies up the method, making it flatter - the error is only used in one spot to determine if we should create a zms instance or not.