Reset the download images preference
Due to a previous mixup, this setting was always being set to the opposite of what it should have been. Fixing that issue means that users who had a previously installed version of wire will have it set off by default when they upgrade, which we don't really want.

Since the preference wasn't really working anyway, I just reset it rather than trying to migrate the value (we need a migration scheme for that...)