Simplify global preferences
Here we now set a preferences signal's value directly in onWire. We also don't need to set the listeners on the UI thread, since the Preferences objects in android are thread safe. Finally, we make sure that the values published to the signal are always pushed to a background thread, since they'll most likely come from the UI